+++ Source/Krypton Components/Krypton.Toolkit.Utilities/Components/KryptonFileSystemListView/Values/FileSystemListViewValues.cs
@@ -211,7 +211,13 @@
             }
         }
 
-        public override bool IsDefault => throw new NotImplementedException();
+        public override bool IsDefault =>
+            _currentPath == string.Empty &&
+            _showFiles &&
+            !_showHiddenFiles &&
+            !_showSystemFiles &&
+            _fileFilter == "*.*" &&
+            !_useLargeIcons;
 
         /// <summary>
         /// Returns a string representation of this object.