+++ Source/Krypton Components/Krypton.Toolkit.Utilities/Components/KryptonFileSystemTreeView/Values/SystemTreeViewValues.cs
@@ -147,7 +147,12 @@ public class SystemTreeViewValues : Storage
         }
     }
 
-    public override bool IsDefault => throw new NotImplementedException();
+    public override bool IsDefault =>
+        _showFiles &&
+        !_showHiddenFiles &&
+        !_showSystemFiles &&
+        _fileFilter == "*.*" &&
+        !_useLargeIcons;
 
     /// <summary>
     /// Returns a string representation of this object.