+++ Source/Krypton Components/Krypton.Toolkit/Values/SystemTreeViewValues.cs
@@ -146,7 +146,12 @@ internal class SystemTreeViewValues : Storage
         }
     }
 
-    public override bool IsDefault => throw new NotImplementedException();
+    public override bool IsDefault =>
+        _showFiles &&
+        !_showHiddenFiles &&
+        !_showSystemFiles &&
+        _fileFilter == "*.*" &&
+        !_useLargeIcons;
 
     /// <summary>
     /// Returns a string representation of this object.